+++ serde_luaq/src/peg_parser.rs
@@ -437,9 +437,7 @@ peg::parser! {
         /// ```
         ///
         /// For more information about Lua type conversion, see [`LuaValue`].
-        ///
-        /// `max_depth` defines the maximum recursion depth for tables.
-        pub rule lua_value(max_depth: usize) -> LuaValue<'input>
+        pub rule lua_value(max_depth: u16) -> LuaValue<'input>
             = (
                 _ "nil" _ { LuaValue::Nil } /
                 _ "true" _ { LuaValue::Boolean(true) } /
@@ -450,7 +448,7 @@ peg::parser! {
                 expected!("Lua value")
             )
 
-        rule table_entry(max_depth: usize) -> LuaTableEntry<'input>
+        rule table_entry(max_depth: u16) -> LuaTableEntry<'input>
             = (
                 // ["foo"]="bar"
                 // [1234]="bar"
@@ -474,10 +472,10 @@ peg::parser! {
                 expected!("Lua table entry")
             )
 
-        rule table_entries(max_depth: usize) -> Vec<LuaTableEntry<'input>>
+        rule table_entries(max_depth: u16) -> Vec<LuaTableEntry<'input>>
             = entries:table_entry(max_depth) ** ([b',' | b';'])
 
-        rule table(max_depth: usize) -> Vec<LuaTableEntry<'input>>
+        rule table(max_depth: u16) -> Vec<LuaTableEntry<'input>>
             =
                 ("{" {?
                     // rust-peg doesn't have a stack limit; workaround based on
@@ -497,11 +495,16 @@ peg::parser! {
                 _
                 "}" { e }
 
-        rule assignment(max_depth: usize) -> (&'input str, LuaValue<'input>)
+        rule assignment(max_depth: u16) -> (&'input str, LuaValue<'input>)
             = i:identifier() _ "=" _ v:lua_value(max_depth) { (i, v) }
 
         /// Parse a Lua script containing variable assignments into a [`Vec`] of
         /// `(&str, LuaValue)`.
+        ///
+        /// For more details about type mapping rules and parameters,
+        /// [see the crate docs][crate#data-types].
+        ///
+        /// ## Example
         ///
         /// ```rust
         /// use serde_luaq::{script, LuaValue};
@@ -516,10 +519,7 @@ peg::parser! {
         /// ```
         ///
         /// For more information about Lua type conversion, see [`LuaValue`].
-        ///
-        /// `max_depth` defines the maximum recursion depth for tables.
-
-        pub rule script(max_depth: usize) -> Vec<(&'input str, LuaValue<'input>)>
+        pub rule script(max_depth: u16) -> Vec<(&'input str, LuaValue<'input>)>
             = (_ a:assignment(max_depth) _ (";" _)* { a })*
 
         /// Parse a Lua `return` stamement into a [`LuaValue`].
@@ -531,9 +531,7 @@ peg::parser! {
         /// ```
         ///
         /// For more information about Lua type conversion, see [`LuaValue`].
-        ///
-        /// `max_depth` defines the maximum recursion depth for tables.
-        pub rule return_statement(max_depth: usize) -> LuaValue<'input>
+        pub rule return_statement(max_depth: u16) -> LuaValue<'input>
             = _ "return" __ v:lua_value(max_depth) _ { v }
     }
 }
